requirements.txt
dj-database-url==0.4.0
Django==1.9.4
django-cors-headers==1.1.0
django-extensions==1.6.1
django-rest-swagger==0.3.5
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
drfdocs==0.0.9
gunicorn==19.4.5
psycopg2==2.6.1
PyJWT==1.4.0
PyYAML==3.11
six==1.10.0
static3==0.6.1
Werkzeug==0.11.4
whitenoise==3.0
